# swagger.yml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Poppy Server API Specification"
...
consumes:
  - "application/json"
  - "application/x-www-form-urlencoded"
produces:
  - "application/json"
schemes:
- "http"
- "https"
securityDefinitions:
  accessToken:
    type: apiKey
    name: Authorization
    in: header
...
paths:
...
  /api/v1/page/page:
    get:
      summary: "Page list for current user"
      tags:
        - "Page"
      responses:
        "200":
          description: ""
      security:
        - accessToken: []
...